requirements.txt
-ipap11helper==0.1
backports.ssl-match-hostname==3.4.0.2
Beaker==1.5.4
blivet==0.61.15.37
Brlapi==0.6.0
cffi==0.8.6
chardet==2.2.1
configobj==4.7.2
configshell-fb==1.1.18
coverage==3.6b3
cryptography==0.8.2
cupshelpers==1.0
decorator==3.4.0
di==0.3
Django==1.8
django-bootstrap3==7.1.0
django-bootstrap3-datetimepicker==2.2.3
django-password-reset==0.9
dnspython==1.12.0
enum34==1.0.4
ethtool==0.8
firstboot==19.5
freeipa==2.0.0a0
fros==1.0
iniparse==0.4
initial-setup==0.3.9.30
ipaddr==2.1.9
ipaplatform==4.2.0
ipapython==4.2.0
IPy==0.75
javapackages==1.0.0
kerberos==1.1
kitchen==1.1.1
kmod==0.1
langtable==0.0.31
lxml==3.2.1
M2Crypto==0.21.1
Mako==0.8.1
MarkupSafe==0.11
netaddr==0.7.5
ntplib==0.3.2
Paste==1.7.5.1
perf==0.1
ply==3.4
policycoreutils-default-encoding==0.1
pyasn1==0.1.6
pycparser==2.14
pycups==1.9.63
pycurl==7.19.0
pygobject==3.14.0
pygpgme==0.3
pyinotify==0.9.4
pykickstart==1.99.66.6
pyliblzma==0.5.3
pyOpenSSL==0.13.1
pyparsing==1.5.6
pyparted==3.9
pysmbc==1.0.13
python-augeas==0.5.0
python-dateutil==1.5
python-default-encoding==0.1
python-dmidecode==3.10.13
python-ldap==2.4.15
python-meh==0.25.2
python-nss==0.16.0
python-yubico==1.2.3
pytz===2012d
pyudev==0.15
pyusb==1.0.0b1
pyxattr==0.5.1
qrcode==5.0.1
rhsm==1.15.4
rtslib-fb==2.1.57
seobject==0.1
sepolicy==1.1
setroubleshoot==1.1
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
SSSDConfig==1.13.0
suds==0.4.1
targetcli-fb===2.1.fb41
Tempita==0.5.1
urlgrabber==3.10
urwid==1.1.1
wheel==0.24.0
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
